extract no apps found to separate component, comment threshold fuse option
